refactor: remove console.logs, move elements files
